Peeking mock call counts and simple Times math
The .Verify + Times.AtLeast is handy, but sometimes you just need to have some more insight on what was going on. Especially, the Times class is lacking support for checking 'incremental' calls. In some cases, it is desirable to not specify how many calls there should be, but rather, how many new calls should occur.

This patch adds a few methods that allow you to get the number of method calls, in a way similar to how the fluent Verify() is normally used.