Delegate lambda currying
In some versions of the Roslyn compiler, lambda backing methods can have an extra parameter to speed up CLR delegate invocations.

We use the valueDel in a few ways, sometimes the sentinel value delegate (no-op) takes in no parameters. In this case, we need to update this to use the new helper method to see if the current value in valueDel accepts no parameters (rather than literally checking to see if its backing method has zero parameters)