Add new method `Times.Matching` to `Times`
This commit adds a new method `Times.Matching(TimesEvaluator)` that
enables client code to pass in custom logic for verifying the expected
number of invocations.

`Times`' public surface remains the same (with the exception of the
added method) but is internally rewritten as a thin wrapper around the
new abstract base class `TimesEvaluator`. Its previous functionality
is off-loaded into two private implementations, `RangeExclusiveBased-
TimesEvaluator` and `RangeInclusiveBasedTimesEvaluator`.