`MockDefaultValueProvider`: Add `Task<T>` support
`MockDefaultValueProvider` does not currently handle the case when it
is invoked for a member returning a `Task<T>` (where `T` is a mockable
type). It will let the underlying `EmptyDefaultValueProvider` produce
a complete `Task` containing `default(T)`.

This commit extends `MockDefaultValueProvider` such that it handles
`Task<T>` itself and produces a complete task containing a mocked `T`.