Reduce likelihood of concurrency bug in `Mock.As<>`
In the preceding commit, some logic was added that might both remove
an item from the `ImplementedInterfaces` list and decrement another
variable (`InternallyImplementedInterfaceCount`). Both should ideally
happen as an atomic operation.

It turns out that the removal is actually irrelevant, and the method
`Mock.As<TInterface>` can be simplified somewhat.