Add failing unit test
Moq currently doesn't allow setting up a method to return `null` via
`Mock.Of`. At current, such a setup seems irrelevant as `null` is
often the default value that doesn't have to be set up explicitly.

However, if we want to allow custom default value providers in the
future, meaning that `null` might no longer be the default value in
all cases, it should be possible to set up `null` explicitly.