Cast expr tree `null` const to right type for !=
Setting up an explicit `null` return value in `Mock.Of` does not work
because a `null` constant in an expression tree is by default of type
`object`, which leads to a type mismatch when Moq uses that constant
to build a new binary expression to do a != comparison. Therefore make
sure to assign a compatible type to `null` const exprs.