Improve method match accuracy in `ExtractProxyCall`
If you're setting up methods in a type hierarchy (e.g. `Base.Method`
and `Derived.Method`), the setup order currently matters, even though
two different methods are set up.

This commits is a first step in making Moq associate the "right" setup
with a proxy method invocation. This will make sure that if there are
several setups that potentially match an invocation, it will attempt
to pick the one for the exact same method.

This is only a start, but probably still not good enough in all
situations (for example, for overridden methods, the distance of the
inheritance relationship is not taken into account).