Fix misreported setup `Times` in `Verify` messages
When `mock.Verify` fails, it lists the mock's setups in the exception
message. However, it reports actual invocation counts as the number
of expected calls set up, which is incorrect.

Moq currently allows only one kind of expected call count setup (both
of which have been obsoleted): `.AtMostOnce()` and `.AtMost(n)`. The
current logic for formatting a setup is moved to `MethodCall` where it
rightly belongs and where the expected call count limit is accessible.