Merge `Moq.csproj` and `Moq.NetCore.csproj`
The separate .NET Framework 4.5 and .NET Standard 1.3 projects are
merged into a single multi-targeting `.csproj`.

Note that this breaks GitLink, which isn't too much of a problem, as
the `.pdb` files aren't currently included in the Moq NuGet package
anyway, but this will have to be addressed in a later commit.