Fix problem with compiler picking wrong overloads
Adding the new generic overload `Returns<TDelegate>(TDelegate)` has
an unfortunate side effect. In some cases, the compiler will pick
the new overload over the preexisting `Returns(TResult)`; for example
when `TResult` is `object` and a call `Returns("string")` is made.

Since we cannot constrain generic type parameters to `Delegate`, we
need to make the new overload non-generic to help the compilers in
picking the right overload.

(There's one more edge case, namely when `TResult` is `Delegate`. The
new overload must forward to the preexisting overload in that case.)