Fix recognition of virtual sealed methods
It turns out that the C# compiler puts the wrong `PropertyInfo` in an
`Expression` for properties in derived classes. Apparently, it always
puts in the `PropertyInfo` corresponding to the property's base dec-
laration instead of the `PropertyInfo` of the accessed type.

This commit adds a clause to the `expression.ToPropertyInfo` extension
method to counteract this peculiar compiler behavior.