Add missing `Setup` methods (except `SetupSet`)
Since `SetupSet`'s parameter cannot be an expression (they may not
contain any assignments), there's nothing to rewrite like with the
other methods; if we stayed true to how Moq usually handles this kind
of expression tree limitation, we'd declare an `Action<TDuck>` param-
eter and then execute it inside a `FluentMockContext`. This doesn't
work either, since we have no instance of `TDuck` to hand to the
delegate.