Add switch for collecting setup source file info
Up until now, Moq has been collecting source file information for each
setups created. This comes at a slight performance penalty. At this
moment, there is a regression in the .NET Framework which means this
penalty becomes gigantic.

This commit turns source file information collection into an opt-in
feature. Towards that end, a `Switches` property is added to `Mock`
and `MockRepository` which allows users to opt in or out of certain
features. Source file info collection is controlled via the switch
`Switches.CollectDiagnosticFileInfoForSetups`.