Extract `InnerMocks` update from mock provider
First, if the default value provider mechanism is to be more generic
in nature, it should be possible for custom providers to create "inner
mocks" without having access to `Mock` internals. This commit makes
it possible for any default value provider to participate in This
mechanism simply by returning a mocked object.

Second, this refactoring is prerequisite to changing the `Provide-
Default` method's signature (so it accepts a `Type` instead of a
`MethodInfo`).