Tests showing problem w/ inner mock verification
It isn't enough for `MockDefaultValueProvider` to correctly wrap a
mocked object in `Task<T>` or `ValueTask<T>`. Produced mocks also have
to be added to the owning / outer mock's `InnerMocks` dictionary in
order to be included in `mock.Verify[All]`.

The tests added in this commit show that this doesn't work yet.