Add `LookupOrFallbackDefaultValueProvider` class
This abstract base class offers the basic functionality required by
providers akin to the empty default value provider. That is, it allows
setting up different value generation strategies for different types,
and it can fall back to a default generation strategy if a request is
made for which no specific strategy is available.

It also handles task types such that values are automatically wrapped
as completed `Task<TResult>` or `ValueTask<TResult>`.

Both `EmptyDefaultValueProvider` and `MockDefaultValueProvider` are
refactored to derive from this type (which incidentally makes them a
tiny bit faster).