Refactor IInterceptStrategy -> InterceptionAspect
Interception "strategy" suggests that only one of several gets chosen
and executed. In truth, all "strategies" are executed. Each one deals
with a specific aspect of a mock method invocation, so rename the type
to reflect this better.

All source files directly related to method interception are put in a
separate sub-folder `Interception` and renamed for better consistency.