Optimize the `Invocation` class (and protocol)
Make both `Method` and `Arguments` non-virtual properties of the base
class. This makes querying those properties faster because delegation
to DynamicProxy is avoided.

Letting the base class handle `Arguments` means that deriving classes
must be explicitly notified about an invocation's end, so that they
can write back modified arguments to by-ref parameters. That is what
the three new `Return` methods are for.

Having Moq explicitly end an invocation makes it possible for Dynamic-
Proxy's `Invocation` subclass to "detach" itself from the underlying
`IInvocation` when is no longer needed, so it can be GC'ed earlier.