Make `SetupAllProperties` semi-lazy
`SetupAllProperties` will still eagerly discover all properties of a
mock's type hierachy, and add setups for all of them, however initial
values for those properties will only be computed when properties are
actually queried for the first time.

This can result in major speedups when `SetupAllProperties` is called
on a mock using `DefaultValue.Mock`.

Making initial value computation lazy has the nice side effect that
type recursion / property loops are no longer a problem (hence the
changes in the unit tests).

Finally, this is a good opportunity to get rid of `SerializableTypes-
ValueProvider`.