Adopt assembly versioning scheme 'major.minor.0.0'
The rationale behind this is that we don't want changes in the patch
version to result in a different strong name; patches should be in-
place updates and shouldn't lead to version conflicts.

This will help reduce the need for assembly binding redirects.

For now, we only do 'major.minor.0.0' instead of 'major.0.0.0' (which
would be even better).

For the next release, we'll have to increase the minor version by 1
to rule out strange-looking binding redirects like these:

    4.0.0.0 - 4.7.145.0  ->  4.0.0.0
    4.0.0.0 - 4.7.145.0  ->  4.7.0.0

In both cases, the following will probably look less suspicious:

    4.0.0.0 - 4.7.145.0  ->  4.8.0.0

Moq 5 will be able to safely make the switch to 'major.0.0.0'.