Fix `CallBase` regression with explicitly implemented interface methods (#558)
* Add failing regression test

which shows that `CallBase` no longer works for explicitly implemented
interface methods (which appear e. g. when `mock.As<TInterface>()` is
used).

* Fix `CallBase` for explicit interface method impls.

While we could just revert 5837c53, that would reintroduce a slow and
hard to understand multi-line `if` condition. Let's instead rewrite
the original conditions such that no superfluous checks are performed,
and such that fast checks are performed first.

(Notably, checking whether a type is an interface is faster than
checking whether it is a class.)

* Update the changelog