Remove superfluous `.StripConversion` helper method (#631)
* Turn `AreSameMethod` into a local function

`AreSameMethod` is only used in one location (inside `VerifyCalls`),
move it there by making it a local function. This in turn makes it
more obvious that `StripConversion` as used in `AreSameMethod` is used
for purely diagnostic purposes here as well.

* Remove `StripConversion` in `AreSameMethod`

Why should `AreSameMethod` be the only location where something like
`StripConversion` is needed? None of the code paths leading up to the
much more important `MethodCall.IsEqualMethodOrOverride` do any cast
removal. So if we don't need this in other places, it's very probably
superfluous here, too.

* Clean-up: Simplify `AreSameMethod`

* Remove other diagnostic uses of `StripConversion`