Rename `IReadOnlyInvocation` to `IInvocation`
and introduce an interface `IInvocationList` instead of using `IRead-
OnlyList<>`. The rationale behind this is that we may want to publish
methods in the future that are not read-only; naming interfaces as
"read-only" would essentially block us from doing so.

Deprecate `mock.ResetCalls()` in favor of `mock.Invocations.Clear()`,
which is more in line with Moq 5's API.