Don't ignore non-public interfaces unconditionally
When mocking a type, Moq currently disregards all its base interfaces
that are declared non-public. This was implemented in 6ad71030b1 to
prevent DynamicProxy from choking on types inaccessible to it.

This produces some collateral damage concerning `CallBase`: even when
`CallBase == true` on the mock, base methods will not be called if
they are declared in a non-public base interface.

Simply excluding non-public base interfaces is too imprecise, because
if there is a suitable `[assembly: InternalsVisibleTo]` custom attri-
bute for DynamicProxy, then even a non-public interface might be
accessible to DynamicProxy, and the error which the type exclusion
supposedly protects against wouldn't occur in the first place.

So instead of just excluding a non-public base interface, let's simply
*ask* DynamicProxy whether it can access it or not.