Remove the `Publish` target
The CI server (AppVeyor) is set up to push packages to NuGet, and this
is how packages should be published. Most users of this build script
wouldn't be able to use `Publish` anyway because they do not know the
NuGet API key required for pushing. So, remove this target lest people
think that they *should* be able to use it.