Replace `dotnet xunit` with `dotnet test`
xUnit 2.4 has abandoned the `dotnet-xunit` CLI tool which we rely on
to run our unit tests. We cannot keep using `dotnet-xunit` if we want
to be able to upgrade xUnit, so let's get rid of it now.

Alternatives considered:

 * `xunit.runner.msbuild` would seem like the obvious choice, but it
   is not cross-platform since it only targets .NET 4.5.

 * `xunit.runner.console` would be the next best thing, since it is
   as feature-complete as `dotnet-xunit` (e.g. it can generate HTML
   and XML test run reports and knows about options like `-noshadow`)
   but it's excruciatingly hard to call, since it comes in various
   versions and we need to map the current `<TargetFramework>` to an
   appropriate command line. Unfortunately, MSBuild multi-targeting
   does not work automatically for `<TargetFrameworks>` and we'd need
   to perform some MSBuild contorsions / recursion to get there.

 * `dotnet test` works well with multi-targeting, but doesn't have all
   options that we've used up until now.

Let's go with the last one for simplicity's sake. If we get complaints
about missing test run report files, we can go back to the more stren-
uous `xunit.runner.console`.