Apply naïve bug fix
When matching a setup against an intercepted method invocation, if the
setup is for an interface method and the invocation is a class method,
we check whether the latter is the implementation for the former; if
not, then the methods don't match.

This is a naïve (because inefficient) bug fix and needs to be improved
but we need to refactor Moq's handling of multi-dot setup expressions
first; see code comment in this commit.