Make `VerifyNoOtherCalls` work with `Verify()`
In order to get `VerifyNoOtherCalls` to interact properly with param-
eterless `Verify[All]()`, we need to check, at verification time, for
all invocations whether they were matched by a (verifiable) setup. If
we have M setups and N invocations, that would be up to M*N checks.

We can do something more efficient.

Moq already checks upon each invocation whether there is a setup that
matches it. So why not cache the result of that check with the invoc-
ation? That allows us to do a linear (N) scan during verification and
flag all setups as verified that were previously matched.

In order to support both `Verify()` and `VerifyAll()`, we need to
distinguish between verifiable setups and regular setups--that's why
the `verified` flag in `Invocation` becomes an enum with that many
options.