Remove impls of uncallable `SetupSet`, `VerifySet`
There used to be overloads of `SetupSet` and `VerifySet` that accepted
two `Expression`s (one selecting the property, the other specifiying a
value for it) instead of a single `Action`. These have since been
marked as `[Obsolete(..., error: true)]`, which means client code that
tries to call them triggers a compile-time error.

Since the methods are in fact no longer callable, we can remove their
implementations and all code that's exclusively used by them.