Move `params` matcher creation into right overload
The logic that creates a matcher for a `params` array parameter would
be better at home in the new `CreateMatcher` overload we created in
the previous refactoring step. Moving it there allows us to get rid of
that `isParams` parameter (i.e. to no longer have a special case) in
the `CreateMatcher` method.