Drop special `mock.Verify(expr)` setup formatting
When producing a verification error message, the parameterised forms
of `mock.Verify` use a divergent formatting for setups: they include
an optional hint about the maximum allowed number of invocations as
set up by the obsolete `.AtMost(n)` and `.AtMostOnce()` setup methods.

That hint does not really serve any purpose, because:

 * The hint is only about the maximum allowed number of invocations.
   When exceeded, an invocation thus set up fails right away and we
   never get to see the error message produced by a subsequent `mock.
   Verify(...)`.

 * If we do get to see the error message, that means that the maximum
   number of allowed invocations hasn't been violated. The real issue
   is just the contrary: the expected invocation (as specified by the
   `Verify` call) has not happened often enough, or an incorrect num-
   ber of times (optionally specified by the `Verify` call). In both
   cases, the maximum number of allowed calls is either unimportant or
   "overridden" by the requirement given to `Verify`.

Let's remove that hint from the formatting, which allows us to finally
abandon the 2nd formatting method of `MethodCall` and just use `.To-
String()` everywhere.