Introduce `SetupWithOutParameterSupport` class
This is needed if we want to convert `SequenceMethodCall` away from
`MethodCall`. `SetupWithOutParameterSupport` captures Moq's slightly
counter-intuitive semantics whereby `out` arguments are read at setup
time to find out what they should be set to during an invocation. This
happens for both regular and sequence setups, but isn't built into
the `Setup` base class, so we need a new in-between class for sequence
setups to derive from--this is it.