Remove unnecessary type check
The check being removed here was introduced by #335, which fixed a
regression by reproducing type matching logic from the .NET reference
source for `Type.GetMethod(string, BindingFlags, Binder, Type[], ..)`.

This check is likely not needed, and we have no test that demonstrates
why it might be needed.

Removing this special check however will allow us to merge the method
it appears in with other similar methods.