Make `VerifyAll` ignore setups from `SetupAllProperties` (#684)
* `VerifyAll` should ignore `SetupAllProperties`

This adds a test demonstrating that `SetupAllProperties` as it is
currently implemented makes `VerifyAll` fail. `SetupAllProperties`
registers a setup for each and every property accessor, requiring the
user to invoke every one of those to make `VerifyAll` pass. This ren-
ders `VerifyAll` rather useless in practice.

* Auto-implemented property invocation no longer required

Those property accessors that are auto-implemented by `SetupAllProper-
ties` no longer need to be invoked in order to pass verification by
`VerifyAll`. (This is easily implemented now that setups verify them-
selves, because for auto-implemented property accessors we have two
dedicated setup types.)

* Update the changelog