Deprecate unsupported verbs after 2nd `Callback`
Non-void method setups allow a second callback after a preceding
`Returns` or `CallBase`. Currently, this is implemented in the fluent
API by reusing the `ICallback` interface, which creates a cycle in
the "flow".

This commit introduces a `ICallbackAfter` and `ICallbackAfterResult`
which together hide or deprecaet the verbs that shouldn't be permitted
after a second `Callback`, hopefully in a way that will cause as
little breakage as possible.

Update the regression tests to reflect which use cases are deprecated
and which ones go away. There's only one of the latter category, and
we can just remove it because there's never been a public Moq release
where it would've been allowed.