Let `SetupSetImpl` changes bubble up into public API
Now that `SetupSetImpl` is no longer generic, the directly calling
methods no longer need to be, either. We can let the non-genericity
"bubble up" into the public API methods where generics surface.

On the side, we get rid of some internal code duplication, as well as
a faulty `Debug.Assert` (which can fail when `SetupSet` is being call-
ed on indexers).