Merge `MethodCallReturn` into `MethodCall`
This will make it possible to remove quite a bit of duplicate code in
the `Mock` class' internal static methods.

Merging `MethodCallReturn` and `MethodCall` could easily make the re-
sulting class heavier for void methods. Because of this, we merge some
fields as well. The net result should be a slightly higher memory
footprint for void methods, and a smaller memory footprint for non-
void methods.