Merge `MethodCallReturn` into `MethodCall` (#717)
* Merge `MethodCallReturn` into `MethodCall`

This will make it possible to remove quite a bit of duplicate code in
the `Mock` class' internal static methods.

Merging `MethodCallReturn` and `MethodCall` could easily make the re-
sulting class heavier for void methods. Because of this, we merge some
fields as well. The net result should be a slightly higher memory
footprint for void methods, and a smaller memory footprint for non-
void methods.

* Reduce footprint of `MethodCall` some more

The setup functions `.AtMostOnce()` and `.AtMost(n)` are deprecated
and probably used only sparingly in practice. This commit offloads all
state associated with these methods into a separate object.

* Remove redundant `MethodCall.FailMessage` setter