Check `IsSpecialName` for properties & indexers
This is a follow-up to #712 and #713, where we added tests documenting
F# and COM events not being marked with the IL `specialname` flag. But
as it turns out, F# and COM *properties and indexer* accessor methods
are marked as `specialname`, which we document here using unit tests.
Finally, we bring back checks for `method.IsSpecialName` for property
and indexer accessor method recognition.