Drop `System.ValueTuple` NuGet dependency
People are still experiencing assembly versioning problems on the .NET
Framework because of the NuGet packages `System.Threading.Tasks.Exten-
sions` and `System.ValueTuple` referenced by Moq.

It is fairly trivial to get rid of the latter dependency... so let's
get rid of it! (It means we can no longer use tuple syntax everywhere
in Moq's code base, but as we haven't made excessive use of it so far,
that's perhaps a small price to pay.)