Merge `HandleTracking` into `RecordInvocation` (#723)
The first of these two interception aspects basically has the same
purpose as the latter: recording that an invocation happened. For this
reason, it would make sense to do combine the former into the latter,
and thus to speed up the interception pipeline.

There's one possible BUT against moving the logic of `HandleTracking`
into `RecordInvocation`: In `RecordInvocation`, we no longer have the
chance to record invocations of "well-known methods" (such as those
defined by `System.Object`) because the interception pipeline might
already have stopped at that point.

Fortunately, this doesn't matter as we are typically not interested in
those well-known methods when making use of `AmbientObserver`: they
don't have signatures that allow chaining or assignment, nor are they
matchers.