Add failing unit test
One of the tests being added here fails because a configured mock is
used as an argument, which Moq executes to find out whether that argu-
ment is a matcher. This is done by means of the `AmbientObserver`,
which has the unfortunate side effect of disabling the argument mock's
setups. Therefore, the expression `data.Id` in `data.Id.Value` will
return `default(Guid?)` instead of the set up value, causing a null
reference exception.