Fix stubbed properties & break `VerifySet`
Fixing stubbed properties involves *not* creating two setups for stub-
bed property getters, but letting `AutoImplementedPropertyGetterSetup`
act as a `FixedReturnValueSetup`. In order to keep the class hierarchy
shallow, we introduce a role interface `IReturnValueSetup` for that
purpose. Inner mock retrieval is then bent a little to use this new
role interface instead of only `FixedReturnValueSetup`.

However, this breaks `VerifySet`. Let's fix that in a separate commit.