Fix regression by avoiding double setups
`FluentMockVisitor` (which is responsible for setting up a full graph
of of mocks for multi-dot expressions) has registered both an `Inner-
Mocks` as well as a setup. The latter could keep track of arguments,
matchers, etc., while the former was used to disover inner mocks. But
now that we've replaced `InnerMocks` with a setup type, we create two
setups, and the latter one overrides the former.

We solve this issue by including argument expression information when
we register a `FixedReturnValueSetup`. That way, the other setup is no
longer required.