Initial impl for LINQ expression reconstructor
Moq needs to work around LINQ expression tree limitations of the .NET
compilers by using delegates instead in some places (e.g. `SetupSet`).
The current approach for turning those back into setups is not ideal.

This commit adds a new component to Moq: the `ExpressionReconstructor`
abstract type, along with an early version of a concrete implementa-
tion (`ActionObserver`). The latter is based on a similar principle as
`AmbientObserver`, but it is completely divorced from the `Mock` type.
It uses an independent interceptor type whose sole raison d'être is
to record invocations and to proxy return types.

This is what we'll be able to do with it:

 1. Eventually install it in `SetupSet`, `VerifySet` and a few other
    places so that these can operate on LINQ expression trees, like
    the rest of the Moq API.

 2. Then remove those code paths from the `Mock` interception pipeline
    that are specific to `AmbientObserver` and performed the same
    function that `ExpressionReconstructor` can do in a more focused
    manner.

These changes might also give us easier-to-understand code and
possibly better runtime performance.