Change "default values" slot selection strategy
So far, matchers only got placed where a parameter received its own
default value. This won't work in a case such as this:

    object Property { get; set; }

    obj => obj.Property = It.IsAny<int>();

because this will invoke `set_Property(object)` with the value 0, i.e.
the default value of the matcher but not of the property type.

If we change the matcher parameter selection rule such that parameters
get chosen whose argument is not equal to a *matcher's* default value.