`Setup[Get|Sequence]`: Migrate to recursive setup
This changes the `Setup`, `SetupGet`, and `SetupSequence` methods to
work in a recursive manner. This is achieved through the new `Split`
method for expressions that we introduced in the previous commit.

(Incidentally, this gets rid of some long-standing code duplication in
the `Mock` class.)

This new approach cannot yet work for `SetupSet`, as that method uses
delegates instead of expressions. Before we can convert `SetupSet` too
we'll need to introduce a new component for reconstructing expressions
from delegates. (This will be added in a future topic branch.)