Convert `SetupSet`, `VerifySet` to expressions
Using the recently added `ExpressionReconstructor` for `SetupSet` and
`VerifySet` means they can now operate on expressions, too; and they
can plug into the new recursive setup/verification algorithm like the
other API methods.

However, some additional work is needed here to get custom matchers
working as expected. It would be inaccurate of `ExpressionReconstruc-
tor` to just embed a custom matcher's `RenderExpression` into the ex-
pression, as that property is only meant to be used for diagnostic
purposes (e.g. in error messages). Instead, the `Match` itself should
be embedded in the expression. Fortunately there's already `MatchEx-
pression` for just that purpose; it just needs to be improved a little
along with `ExpressionComparer` and `ExpressionStringBuilder`. An ex-
pression containing `MatchExpression`s should never be compiled and
executed. To ensure that, we declare this node type as irreducible.