Add more precise inner mock setups
Now that we're more picky when choosing inner mock setups, we should
also be more accurate when adding them. Otherwise, the only matches
we'll ever get will be if all arguments use `It.IsAny<TArg>`, as that
is the kind of inner mock setup generated by `DefaultValue.Mock`.

This commit makes the tests pass that originally failed (_1 and _2),
but it now breaks those using `It.IsAny` for verification (_3).

This is an important breakage as it demonstrates a fundamental problem
in the current recursive verification approach: In a verification call
for `m => m.X.Y`, the inner mock `Y` to proceed to from `X` should not
be chosen from `X`'s setups, but from the return values of all invoc-
ations matching `X` (and there may be several different ones)!