Make `stringBuilder.AppendValueOf` safer
...by only showing the first 10 items of enumerables for certain safe
collection types (arrays and `List<T>`). It cannot be generally assum-
ed that `IEnumerable` actually work.

I cannot easily add the test case provided in #741 to the regression
test suite as it relies on EF Core, and the test project already uses
EF, so we'd likely have type conflicts if both of them are referenced.