Add tests for `default(Times)`
Specify that `default(Times)` should be equivalent to `Times.AtLeast-
Once()`, which is the implicit assumption of most public API method
overloads that don't explicitly ask for a `Times` instance.

While the use of `default(Times)` isn't encouraged, this still needs
to be made to work as `Times` is a struct, for which sensible default
values are a must. (In previous versions of Moq, `default(Times)` will
cause `NullReferenceException`s sooner or later; not too great.)