More consistent default return values w/ loose mocks
This fixes a long-standing inconsistency in how Moq produces default
return values. Given a loose mock for a type which has some member
`TResult M()`:

 * If there's no setup for `M`, the return value is determined by
   `mock.CallBase` or, if it is not set, by `mock.DefaultValue-
    [Provider]`.

 * If there is a setup for `M`, which has either a `.Returns(...)` or
   a `.CallBase()` clause, then the return value will be determined by
   that.

 * If there is a setup for `M`, but the setup has neither a
   `.Returns(...)` nor a `.CallBase()` clause, then `default(T)` will
   be returned.

The last one doesn't really fit in with the rest of how Moq handles
default values. While it may be a user mistake not to explicitly
specify the return value for a setup, it cannot hurt to make things
more consistent.

Better consistency in this case means that such "incomplete" setups
(3) will produce the same return value for M as (1).